add feature=host to types
