+++ crates/types/batch/src/lib.rs
@@ -23,3 +23,5 @@ pub use witness::{
     ArchivedLegacyBatchWitness, BatchWitness, Bytes48, LegacyBatchWitness, PointEvalWitness,
     build_point_eval_witness,
 };
+
+pub mod utils;
