+++ crates/types/chunk/Cargo.toml
@@ -24,6 +24,6 @@ types-base = { path = "../base", package = "scroll-zkvm-types-base"}
 
 [features]
 default = []
-# should be enabled at host
+host = ["scroll-compress-ratio"]
 scroll-compress-ratio = ["sbv-primitives/scroll-compress-ratio"]
 openvm = ["sbv-core/scroll-openvm"]
